compile,propagate(gRain)	R Documentation

Compile and propagate functions

Description

compile and propagate are generic functions which invoke particular methods which depend on the class of the first argument

Usage


compile(object, ...)
propagate(object, ...)

Arguments

object 	 An object to be compiled or propagated

... 	 Additional arguments which depends on the class of the object

Value

The value returned depends on the class of the first argument.

Author(s)

Søren Højsgaard, sorenh@agrsci.dk

See Also

compile.grain, propagate.grain

[Package gRain version 0.8.0 Index]