compileCPT,compilePOT(gRain)	R Documentation

Compile conditional probability tables / cliques potentials.

Description

Compile conditional probability tables / cliques potentials as a preprocessing step for creating a graphical independence network

Usage


compileCPT(x)
compilePOT(x)

Arguments

x 	 To compileCPT x is a list of conditional probability tables; to compilePOT, x is a list of clique potentials

Value

compileCPT returns a list of class 'cptspec' compilePOT returns a list of class 'potspec'

Author(s)

~~who you are~~

See Also

extractCPT, extractPOT

[Package gRain version 0.8.2 Index]