gRain-package(gRain)	R Documentation

Overview - the gRain package for graphical independence networks

Description

Probability propagation in graphical independence networks, also known as probabilistic expert systems (which includes Bayesian networks).

Details

This package implements graphical independence networks.

The package is still at an experimental stage.

The main function for building networks is grain.

The function querygrain is used for querying independence networks.

Functions simulate.grain and predict.grain are available.

There is a small vignette which illustrates the use of gRain.

A paper describing gRain is under preparation

Author(s)

Søren Højsgaard <sorenh@agrsci.dk>

Maintainer: Søren Højsgaard <sorenh@agrsci.dk>

[Package gRain version 0.8.0 Index]