Set joint evidence in grain objects

Setting and removing joint evidence in grain objects.

setJEvidence(object, evidence = NULL, propagate = TRUE, details = 0)

retractJEvidence(object, items = NULL, propagate = TRUE, details = 0)

new_jev(ev, levels)

Arguments

object

A "grain" object.

evidence

A list of evidence. Each element is a named array.

propagate

Should evidence be absorbed once entered; defaults to TRUE.

details

Amount of printing; for debugging.

items

Items in the evidence list to be removed. Here, NULL means remove everything, 0 means nothing is removed. Otherwise items is a numeric vector.

ev

A named list.

levels

A named list.

Note

All the joint evidence functionality should be used with great care.

Author

Søren Højsgaard, sorenh@math.aau.dk

Examples

example("example_chest_cpt")
#> 
#> exmp__> yn   <- c("yes", "no")
#> 
#> exmp__> a    <- cpt(~asia, values=c(1,99),levels=yn)
#> 
#> exmp__> t.a  <- cpt(~tub|asia, values=c(5,95,1,99),levels=yn)
#> 
#> exmp__> s    <- cpt(~smoke, values=c(5,5), levels=yn)
#> 
#> exmp__> l.s  <- cpt(~lung|smoke, values=c(1,9,1,99), levels=yn)
#> 
#> exmp__> b.s  <- cpt(~bronc|smoke, values=c(6,4,3,7), levels=yn)
#> 
#> exmp__> e.lt <- cpt(~either|lung:tub,values=c(1,0,1,0,1,0,0,1),levels=yn)
#> 
#> exmp__> x.e  <- cpt(~xray|either, values=c(98,2,5,95), levels=yn)
#> 
#> exmp__> d.be <- cpt(~dysp|bronc:either, values=c(9,1,7,3,8,2,1,9), levels=yn)
#> 
#> exmp__> chest_cpt <- list(a, t.a, s, l.s, b.s, e.lt, x.e, d.be)
#> 
#> exmp__> ## bn <- grain(compile_cpt(chest_cpt))
#> exmp__> 
#> exmp__> 
#> exmp__> 
chest.bn <- grain(compileCPT(chest_cpt))
chest.bn <- compile(chest.bn)

uni <- list(asia = c("yes", "no"), tub = c("yes", "no"),
            smoke = c("yes", "no"), lung = c("yes", "no"),
            bronc = c("yes", "no"), either = c("yes", "no"),
            xray = c("yes", "no"), dysp = c("yes", "no"))

ev <- list(tabNew("asia", levels=uni, values=c(1,0)),
           tabNew("dysp", levels=uni, values=c(1,0)),
           tabNew(c("dysp","bronc"), levels=uni, values=c(.1, .2, .9, .8)) )

chest.bn
#> Independence network: Compiled: TRUE Propagated: FALSE Evidence: FALSE 
chest.bn2 <- setJEvidence(chest.bn, evidence=ev)
chest.bn2
#> Independence network: Compiled: TRUE Propagated: TRUE Evidence: TRUE 
getEvidence(chest.bn2)
#> [[1]]
#> asia
#> yes  no 
#>   1   0 
#> 
#> [[2]]
#> dysp
#> yes  no 
#>   1   0 
#> 
#> [[3]]
#>      bronc
#> dysp  yes  no
#>   yes 0.1 0.9
#>   no  0.2 0.8
#> 
#> attr(,"class")
#> [1] "grain_joint_evidence"

# Notice: The evidence is defined on (subsets of) cliques of the junction tree
# and therefore evidence can readily be absorbed:
getgrain(chest.bn, "rip")$cliques |> str()
#> List of 6
#>  $ : chr [1:2] "asia" "tub"
#>  $ : chr [1:3] "lung" "either" "tub"
#>  $ : chr [1:3] "lung" "either" "bronc"
#>  $ : chr [1:3] "smoke" "lung" "bronc"
#>  $ : chr [1:3] "bronc" "either" "dysp"
#>  $ : chr [1:2] "xray" "either"

# On the other hand, below is evidence which is not defined cliques
# of the junction tree and therefore evidence can not easily be
# absorbed.  Hence this will fail:

if (FALSE) { # \dontrun{
ev.fail <- list(tabNew(c("dysp","smoke"), levels=uni, values=c(.1, .2, .9, .8)) )
setJEvidence(chest.bn, evidence=ev.fail)
} # }

## Evidence can be removed with

retractJEvidence(chest.bn2)      ## All evidence removed.
#> Independence network: Compiled: TRUE Propagated: TRUE Evidence: FALSE 
retractJEvidence(chest.bn2, 0)   ## No evidence removed.
#> Independence network: Compiled: TRUE Propagated: TRUE Evidence: TRUE 
retractJEvidence(chest.bn2, 1:2) ## Evidence items 1 and 2 are removed.
#> Independence network: Compiled: TRUE Propagated: TRUE Evidence: TRUE 

# Setting additional joint evidence to an object where joint
# evidence already is set will cause an error. Hence this will fail:

if (FALSE) { # \dontrun{
  ev2 <- list(smoke="yes")
  setJEvidence(chest.bn2, evidence=ev2)
} # }

## Instead we can do
new.ev <- c(getEvidence(chest.bn2), list(smoke="yes"))
chest.bn
#> Independence network: Compiled: TRUE Propagated: FALSE Evidence: FALSE 
setJEvidence(chest.bn, evidence=new.ev)
#> Independence network: Compiled: TRUE Propagated: TRUE Evidence: TRUE 

## Create joint evidence object:
yn <- c("yes", "no")
db <- tabNew(c("dysp", "bronc"), list(dysp=yn, bronc=yn), values=c(.1, .2, .9, .8))
db
#>      bronc
#> dysp  yes  no
#>   yes 0.1 0.9
#>   no  0.2 0.8
ev   <- list(asia=c(1, 0), dysp="yes", db)

jevi <- new_jev(ev, levels=uni)
jevi
#> [[1]]
#> asia
#> yes  no 
#>   1   0 
#> 
#> [[2]]
#> dysp
#> yes  no 
#>   1   0 
#> 
#> [[3]]
#>      bronc
#> dysp  yes  no
#>   yes 0.1 0.9
#>   no  0.2 0.8
#> 
#> attr(,"class")
#> [1] "grain_joint_evidence"

chest.bn3 <- setJEvidence(chest.bn, evidence=jevi)
getEvidence(chest.bn3)
#> [[1]]
#> asia
#> yes  no 
#>   1   0 
#> 
#> [[2]]
#> dysp
#> yes  no 
#>   1   0 
#> 
#> [[3]]
#>      bronc
#> dysp  yes  no
#>   yes 0.1 0.9
#>   no  0.2 0.8
#> 
#> attr(,"class")
#> [1] "grain_joint_evidence"