Make predictions from a probabilistic network

Makes predictions (either as the most likely state or as the conditional distributions) of variables conditional on finding (evidence) on other variables in an independence network.

# S3 method for class 'grain'
predict(
  object,
  response,
  predictors = setdiff(names(newdata), response),
  newdata,
  type = "class",
  ...
)

Arguments

object

A grain object

response

A vector of response variables to make predictions on

predictors

A vector of predictor variables to make predictions from. Defaults to all variables that are note responses.

newdata

A data frame

type

If "class", the most probable class is returned; if "distribution" the conditional distribution is returned.

...

Not used

Value

A list with components

pred

A list with the predictions

pFinding

A vector with the probability of the finding (evidence) on which the prediction is based

References

Søren Højsgaard (2012). Graphical Independence Networks with the gRain Package for R. Journal of Statistical Software, 46(10), 1-26. https://www.jstatsoft.org/v46/i10/.

See also

Author

Søren Højsgaard, sorenh@math.aau.dk

Examples

example("example_chest_cpt")
#> 
#> exmp__> yn   <- c("yes", "no")
#> 
#> exmp__> a    <- cpt(~asia, values=c(1,99),levels=yn)
#> 
#> exmp__> t.a  <- cpt(~tub|asia, values=c(5,95,1,99),levels=yn)
#> 
#> exmp__> s    <- cpt(~smoke, values=c(5,5), levels=yn)
#> 
#> exmp__> l.s  <- cpt(~lung|smoke, values=c(1,9,1,99), levels=yn)
#> 
#> exmp__> b.s  <- cpt(~bronc|smoke, values=c(6,4,3,7), levels=yn)
#> 
#> exmp__> e.lt <- cpt(~either|lung:tub,values=c(1,0,1,0,1,0,0,1),levels=yn)
#> 
#> exmp__> x.e  <- cpt(~xray|either, values=c(98,2,5,95), levels=yn)
#> 
#> exmp__> d.be <- cpt(~dysp|bronc:either, values=c(9,1,7,3,8,2,1,9), levels=yn)
#> 
#> exmp__> chest_cpt <- list(a, t.a, s, l.s, b.s, e.lt, x.e, d.be)
#> 
#> exmp__> ## bn <- grain(compile_cpt(chest_cpt))
#> exmp__> 
#> exmp__> 
#> exmp__> 
data(chestSim500)

chest.bn <- grain(compileCPT(chest_cpt))
nd <- chestSim500[1:4]

predict(chest.bn, response="bronc", newdata=nd)
#> $pred
#> $pred$bronc
#>   [1] "no"  "no"  "yes" "yes" "yes" "no"  "yes" "no"  "no"  "no"  "no"  "no" 
#>  [13] "yes" "no"  "yes" "yes" "no"  "yes" "no"  "no"  "no"  "yes" "no"  "no" 
#>  [25] "yes" "yes" "yes" "no"  "no"  "yes" "yes" "yes" "no"  "no"  "no"  "yes"
#>  [37] "no"  "no"  "no"  "yes" "yes" "no"  "no"  "no"  "yes" "no"  "no"  "yes"
#>  [49] "no"  "no"  "no"  "yes" "no"  "yes" "no"  "no"  "yes" "no"  "no"  "no" 
#>  [61] "no"  "no"  "no"  "yes" "yes" "yes" "no"  "no"  "yes" "no"  "yes" "no" 
#>  [73] "yes" "no"  "no"  "no"  "no"  "no"  "no"  "no"  "no"  "no"  "no"  "no" 
#>  [85] "no"  "yes" "yes" "yes" "yes" "no"  "yes" "no"  "no"  "yes" "no"  "no" 
#>  [97] "no"  "no"  "no"  "yes" "yes" "yes" "yes" "yes" "yes" "yes" "yes" "no" 
#> [109] "no"  "yes" "no"  "no"  "no"  "no"  "yes" "no"  "yes" "yes" "no"  "no" 
#> [121] "yes" "no"  "yes" "no"  "no"  "yes" "yes" "no"  "yes" "yes" "no"  "no" 
#> [133] "yes" "no"  "yes" "yes" "yes" "no"  "yes" "yes" "no"  "no"  "yes" "no" 
#> [145] "no"  "no"  "yes" "no"  "yes" "no"  "no"  "no"  "yes" "no"  "no"  "yes"
#> [157] "no"  "yes" "yes" "no"  "no"  "yes" "yes" "no"  "yes" "no"  "no"  "no" 
#> [169] "yes" "yes" "no"  "no"  "yes" "no"  "yes" "yes" "yes" "no"  "no"  "yes"
#> [181] "yes" "no"  "yes" "yes" "yes" "yes" "no"  "no"  "yes" "yes" "yes" "no" 
#> [193] "no"  "yes" "no"  "no"  "yes" "yes" "no"  "no"  "no"  "no"  "yes" "no" 
#> [205] "no"  "yes" "no"  "yes" "yes" "no"  "yes" "yes" "no"  "yes" "no"  "yes"
#> [217] "yes" "no"  "no"  "no"  "yes" "no"  "yes" "no"  "yes" "yes" "yes" "no" 
#> [229] "no"  "yes" "yes" "yes" "yes" "no"  "no"  "yes" "no"  "no"  "no"  "no" 
#> [241] "yes" "yes" "no"  "yes" "no"  "no"  "yes" "no"  "no"  "yes" "yes" "yes"
#> [253] "yes" "yes" "no"  "yes" "no"  "yes" "no"  "no"  "yes" "yes" "no"  "no" 
#> [265] "no"  "no"  "no"  "no"  "no"  "no"  "no"  "yes" "no"  "yes" "no"  "yes"
#> [277] "yes" "no"  "no"  "yes" "no"  "yes" "no"  "no"  "yes" "yes" "yes" "no" 
#> [289] "no"  "no"  "no"  "yes" "yes" "no"  "no"  "yes" "yes" "no"  "yes" "no" 
#> [301] "no"  "no"  "no"  "no"  "yes" "yes" "yes" "yes" "no"  "yes" "no"  "no" 
#> [313] "yes" "yes" "yes" "no"  "no"  "no"  "yes" "no"  "no"  "no"  "yes" "yes"
#> [325] "no"  "yes" "yes" "no"  "yes" "yes" "yes" "no"  "no"  "no"  "yes" "no" 
#> [337] "no"  "yes" "yes" "no"  "no"  "yes" "no"  "no"  "yes" "no"  "yes" "yes"
#> [349] "yes" "no"  "yes" "no"  "no"  "yes" "no"  "yes" "yes" "no"  "yes" "yes"
#> [361] "no"  "no"  "yes" "no"  "no"  "yes" "no"  "no"  "yes" "yes" "yes" "yes"
#> [373] "no"  "yes" "no"  "no"  "yes" "yes" "no"  "yes" "no"  "yes" "yes" "yes"
#> [385] "yes" "no"  "no"  "yes" "no"  "yes" "yes" "no"  "no"  "yes" "yes" "no" 
#> [397] "yes" "yes" "no"  "yes" "no"  "no"  "no"  "yes" "yes" "no"  "yes" "yes"
#> [409] "no"  "no"  "yes" "yes" "no"  "no"  "yes" "yes" "no"  "yes" "yes" "yes"
#> [421] "yes" "yes" "yes" "no"  "no"  "no"  "yes" "yes" "yes" "yes" "no"  "no" 
#> [433] "no"  "no"  "yes" "no"  "no"  "yes" "yes" "no"  "yes" "no"  "no"  "yes"
#> [445] "yes" "no"  "yes" "yes" "no"  "no"  "no"  "yes" "yes" "no"  "no"  "no" 
#> [457] "no"  "yes" "yes" "no"  "no"  "yes" "yes" "no"  "yes" "yes" "yes" "yes"
#> [469] "no"  "no"  "no"  "yes" "yes" "yes" "no"  "no"  "yes" "no"  "yes" "yes"
#> [481] "yes" "no"  "yes" "yes" "yes" "yes" "yes" "no"  "no"  "no"  "yes" "no" 
#> [493] "yes" "no"  "yes" "yes" "no"  "yes" "no"  "no" 
#> 
#> 
#> $pEvidence
#>   [1] 0.4851495 0.4851495 0.4410450 0.4410450 0.4410450 0.4851495 0.4410450
#>   [8] 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495 0.0490050 0.4851495
#>  [15] 0.4410450 0.4410450 0.4851495 0.4410450 0.4851495 0.4851495 0.4851495
#>  [22] 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450 0.4410450 0.4851495
#>  [29] 0.4851495 0.4410450 0.4410450 0.4410450 0.4851495 0.4851495 0.4851495
#>  [36] 0.4410450 0.4851495 0.0049005 0.4851495 0.4410450 0.4410450 0.4851495
#>  [43] 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495
#>  [50] 0.4851495 0.4851495 0.4410450 0.4851495 0.4410450 0.4851495 0.4851495
#>  [57] 0.4410450 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495
#>  [64] 0.4410450 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495
#>  [71] 0.4410450 0.4851495 0.4410450 0.4851495 0.4851495 0.4851495 0.0049005
#>  [78] 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495
#>  [85] 0.0049005 0.4410450 0.4410450 0.4410450 0.4410450 0.4851495 0.4410450
#>  [92] 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495 0.4851495 0.4851495
#>  [99] 0.4851495 0.4410450 0.0490050 0.4410450 0.4410450 0.4410450 0.4410450
#> [106] 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495
#> [113] 0.4851495 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495
#> [120] 0.4851495 0.4410450 0.4851495 0.4410450 0.4851495 0.4851495 0.4410450
#> [127] 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450
#> [134] 0.4851495 0.0490050 0.4410450 0.0490050 0.4851495 0.4410450 0.4410450
#> [141] 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495 0.4851495 0.4410450
#> [148] 0.0049005 0.4410450 0.4851495 0.4851495 0.4851495 0.4410450 0.4851495
#> [155] 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495
#> [162] 0.4410450 0.4410450 0.4851495 0.4410450 0.4851495 0.4851495 0.4851495
#> [169] 0.4410450 0.0490050 0.0000495 0.4851495 0.4410450 0.4851495 0.4410450
#> [176] 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450 0.4851495
#> [183] 0.4410450 0.4410450 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450
#> [190] 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495
#> [197] 0.4410450 0.4410450 0.4851495 0.4851495 0.4851495 0.4851495 0.0490050
#> [204] 0.4851495 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495
#> [211] 0.4410450 0.4410450 0.4851495 0.0490050 0.4851495 0.4410450 0.4410450
#> [218] 0.4851495 0.4851495 0.4851495 0.4410450 0.4851495 0.4410450 0.4851495
#> [225] 0.4410450 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450
#> [232] 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495
#> [239] 0.4851495 0.4851495 0.4410450 0.4410450 0.4851495 0.4410450 0.4851495
#> [246] 0.4851495 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450 0.4410450
#> [253] 0.4410450 0.4410450 0.4851495 0.4410450 0.4851495 0.4410450 0.4851495
#> [260] 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495 0.4851495 0.4851495
#> [267] 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495 0.0490050 0.4851495
#> [274] 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495 0.0490050
#> [281] 0.4851495 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450 0.4410450
#> [288] 0.4851495 0.4851495 0.4851495 0.4851495 0.4410450 0.4410450 0.4851495
#> [295] 0.4851495 0.4410450 0.4410450 0.4851495 0.4410450 0.4851495 0.0049005
#> [302] 0.4851495 0.4851495 0.4851495 0.4410450 0.4410450 0.0490050 0.4410450
#> [309] 0.4851495 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450 0.4410450
#> [316] 0.4851495 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495 0.4851495
#> [323] 0.4410450 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495 0.4410450
#> [330] 0.4410450 0.4410450 0.4851495 0.4851495 0.0049005 0.4410450 0.4851495
#> [337] 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495
#> [344] 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4410450 0.4851495
#> [351] 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450
#> [358] 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495
#> [365] 0.4851495 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450 0.4410450
#> [372] 0.4410450 0.4851495 0.0490050 0.4851495 0.4851495 0.4410450 0.0490050
#> [379] 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4410450 0.4410450
#> [386] 0.4851495 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495
#> [393] 0.4851495 0.4410450 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495
#> [400] 0.4410450 0.4851495 0.4851495 0.4851495 0.4410450 0.0490050 0.4851495
#> [407] 0.0042750 0.4410450 0.4851495 0.4851495 0.4410450 0.0490050 0.4851495
#> [414] 0.4851495 0.0042750 0.4410450 0.4851495 0.4410450 0.4410450 0.4410450
#> [421] 0.4410450 0.4410450 0.4410450 0.4851495 0.4851495 0.4851495 0.4410450
#> [428] 0.4410450 0.0490050 0.4410450 0.4851495 0.4851495 0.4851495 0.4851495
#> [435] 0.4410450 0.4851495 0.4851495 0.0490050 0.4410450 0.4851495 0.4410450
#> [442] 0.4851495 0.4851495 0.4410450 0.0490050 0.4851495 0.4410450 0.4410450
#> [449] 0.4851495 0.4851495 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495
#> [456] 0.0049005 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450
#> [463] 0.4410450 0.4851495 0.4410450 0.4410450 0.4410450 0.4410450 0.4851495
#> [470] 0.4851495 0.0049005 0.0490050 0.4410450 0.4410450 0.4851495 0.4851495
#> [477] 0.0002250 0.4851495 0.4410450 0.4410450 0.4410450 0.4851495 0.4410450
#> [484] 0.0490050 0.4410450 0.4410450 0.4410450 0.4851495 0.4851495 0.4851495
#> [491] 0.4410450 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495
#> [498] 0.4410450 0.4851495 0.4851495
#> 
predict(chest.bn, response="bronc", newdata=nd, type="distribution")
#> $pred
#> $pred$bronc
#>        yes  no
#>   [1,] 0.3 0.7
#>   [2,] 0.3 0.7
#>   [3,] 0.6 0.4
#>   [4,] 0.6 0.4
#>   [5,] 0.6 0.4
#>   [6,] 0.3 0.7
#>   [7,] 0.6 0.4
#>   [8,] 0.3 0.7
#>   [9,] 0.3 0.7
#>  [10,] 0.3 0.7
#>  [11,] 0.3 0.7
#>  [12,] 0.3 0.7
#>  [13,] 0.6 0.4
#>  [14,] 0.3 0.7
#>  [15,] 0.6 0.4
#>  [16,] 0.6 0.4
#>  [17,] 0.3 0.7
#>  [18,] 0.6 0.4
#>  [19,] 0.3 0.7
#>  [20,] 0.3 0.7
#>  [21,] 0.3 0.7
#>  [22,] 0.6 0.4
#>  [23,] 0.3 0.7
#>  [24,] 0.3 0.7
#>  [25,] 0.6 0.4
#>  [26,] 0.6 0.4
#>  [27,] 0.6 0.4
#>  [28,] 0.3 0.7
#>  [29,] 0.3 0.7
#>  [30,] 0.6 0.4
#>  [31,] 0.6 0.4
#>  [32,] 0.6 0.4
#>  [33,] 0.3 0.7
#>  [34,] 0.3 0.7
#>  [35,] 0.3 0.7
#>  [36,] 0.6 0.4
#>  [37,] 0.3 0.7
#>  [38,] 0.3 0.7
#>  [39,] 0.3 0.7
#>  [40,] 0.6 0.4
#>  [41,] 0.6 0.4
#>  [42,] 0.3 0.7
#>  [43,] 0.3 0.7
#>  [44,] 0.3 0.7
#>  [45,] 0.6 0.4
#>  [46,] 0.3 0.7
#>  [47,] 0.3 0.7
#>  [48,] 0.6 0.4
#>  [49,] 0.3 0.7
#>  [50,] 0.3 0.7
#>  [51,] 0.3 0.7
#>  [52,] 0.6 0.4
#>  [53,] 0.3 0.7
#>  [54,] 0.6 0.4
#>  [55,] 0.3 0.7
#>  [56,] 0.3 0.7
#>  [57,] 0.6 0.4
#>  [58,] 0.3 0.7
#>  [59,] 0.3 0.7
#>  [60,] 0.3 0.7
#>  [61,] 0.3 0.7
#>  [62,] 0.3 0.7
#>  [63,] 0.3 0.7
#>  [64,] 0.6 0.4
#>  [65,] 0.6 0.4
#>  [66,] 0.6 0.4
#>  [67,] 0.3 0.7
#>  [68,] 0.3 0.7
#>  [69,] 0.6 0.4
#>  [70,] 0.3 0.7
#>  [71,] 0.6 0.4
#>  [72,] 0.3 0.7
#>  [73,] 0.6 0.4
#>  [74,] 0.3 0.7
#>  [75,] 0.3 0.7
#>  [76,] 0.3 0.7
#>  [77,] 0.3 0.7
#>  [78,] 0.3 0.7
#>  [79,] 0.3 0.7
#>  [80,] 0.3 0.7
#>  [81,] 0.3 0.7
#>  [82,] 0.3 0.7
#>  [83,] 0.3 0.7
#>  [84,] 0.3 0.7
#>  [85,] 0.3 0.7
#>  [86,] 0.6 0.4
#>  [87,] 0.6 0.4
#>  [88,] 0.6 0.4
#>  [89,] 0.6 0.4
#>  [90,] 0.3 0.7
#>  [91,] 0.6 0.4
#>  [92,] 0.3 0.7
#>  [93,] 0.3 0.7
#>  [94,] 0.6 0.4
#>  [95,] 0.3 0.7
#>  [96,] 0.3 0.7
#>  [97,] 0.3 0.7
#>  [98,] 0.3 0.7
#>  [99,] 0.3 0.7
#> [100,] 0.6 0.4
#> [101,] 0.6 0.4
#> [102,] 0.6 0.4
#> [103,] 0.6 0.4
#> [104,] 0.6 0.4
#> [105,] 0.6 0.4
#> [106,] 0.6 0.4
#> [107,] 0.6 0.4
#> [108,] 0.3 0.7
#> [109,] 0.3 0.7
#> [110,] 0.6 0.4
#> [111,] 0.3 0.7
#> [112,] 0.3 0.7
#> [113,] 0.3 0.7
#> [114,] 0.3 0.7
#> [115,] 0.6 0.4
#> [116,] 0.3 0.7
#> [117,] 0.6 0.4
#> [118,] 0.6 0.4
#> [119,] 0.3 0.7
#> [120,] 0.3 0.7
#> [121,] 0.6 0.4
#> [122,] 0.3 0.7
#> [123,] 0.6 0.4
#> [124,] 0.3 0.7
#> [125,] 0.3 0.7
#> [126,] 0.6 0.4
#> [127,] 0.6 0.4
#> [128,] 0.3 0.7
#> [129,] 0.6 0.4
#> [130,] 0.6 0.4
#> [131,] 0.3 0.7
#> [132,] 0.3 0.7
#> [133,] 0.6 0.4
#> [134,] 0.3 0.7
#> [135,] 0.6 0.4
#> [136,] 0.6 0.4
#> [137,] 0.6 0.4
#> [138,] 0.3 0.7
#> [139,] 0.6 0.4
#> [140,] 0.6 0.4
#> [141,] 0.3 0.7
#> [142,] 0.3 0.7
#> [143,] 0.6 0.4
#> [144,] 0.3 0.7
#> [145,] 0.3 0.7
#> [146,] 0.3 0.7
#> [147,] 0.6 0.4
#> [148,] 0.3 0.7
#> [149,] 0.6 0.4
#> [150,] 0.3 0.7
#> [151,] 0.3 0.7
#> [152,] 0.3 0.7
#> [153,] 0.6 0.4
#> [154,] 0.3 0.7
#> [155,] 0.3 0.7
#> [156,] 0.6 0.4
#> [157,] 0.3 0.7
#> [158,] 0.6 0.4
#> [159,] 0.6 0.4
#> [160,] 0.3 0.7
#> [161,] 0.3 0.7
#> [162,] 0.6 0.4
#> [163,] 0.6 0.4
#> [164,] 0.3 0.7
#> [165,] 0.6 0.4
#> [166,] 0.3 0.7
#> [167,] 0.3 0.7
#> [168,] 0.3 0.7
#> [169,] 0.6 0.4
#> [170,] 0.6 0.4
#> [171,] 0.3 0.7
#> [172,] 0.3 0.7
#> [173,] 0.6 0.4
#> [174,] 0.3 0.7
#> [175,] 0.6 0.4
#> [176,] 0.6 0.4
#> [177,] 0.6 0.4
#> [178,] 0.3 0.7
#> [179,] 0.3 0.7
#> [180,] 0.6 0.4
#> [181,] 0.6 0.4
#> [182,] 0.3 0.7
#> [183,] 0.6 0.4
#> [184,] 0.6 0.4
#> [185,] 0.6 0.4
#> [186,] 0.6 0.4
#> [187,] 0.3 0.7
#> [188,] 0.3 0.7
#> [189,] 0.6 0.4
#> [190,] 0.6 0.4
#> [191,] 0.6 0.4
#> [192,] 0.3 0.7
#> [193,] 0.3 0.7
#> [194,] 0.6 0.4
#> [195,] 0.3 0.7
#> [196,] 0.3 0.7
#> [197,] 0.6 0.4
#> [198,] 0.6 0.4
#> [199,] 0.3 0.7
#> [200,] 0.3 0.7
#> [201,] 0.3 0.7
#> [202,] 0.3 0.7
#> [203,] 0.6 0.4
#> [204,] 0.3 0.7
#> [205,] 0.3 0.7
#> [206,] 0.6 0.4
#> [207,] 0.3 0.7
#> [208,] 0.6 0.4
#> [209,] 0.6 0.4
#> [210,] 0.3 0.7
#> [211,] 0.6 0.4
#> [212,] 0.6 0.4
#> [213,] 0.3 0.7
#> [214,] 0.6 0.4
#> [215,] 0.3 0.7
#> [216,] 0.6 0.4
#> [217,] 0.6 0.4
#> [218,] 0.3 0.7
#> [219,] 0.3 0.7
#> [220,] 0.3 0.7
#> [221,] 0.6 0.4
#> [222,] 0.3 0.7
#> [223,] 0.6 0.4
#> [224,] 0.3 0.7
#> [225,] 0.6 0.4
#> [226,] 0.6 0.4
#> [227,] 0.6 0.4
#> [228,] 0.3 0.7
#> [229,] 0.3 0.7
#> [230,] 0.6 0.4
#> [231,] 0.6 0.4
#> [232,] 0.6 0.4
#> [233,] 0.6 0.4
#> [234,] 0.3 0.7
#> [235,] 0.3 0.7
#> [236,] 0.6 0.4
#> [237,] 0.3 0.7
#> [238,] 0.3 0.7
#> [239,] 0.3 0.7
#> [240,] 0.3 0.7
#> [241,] 0.6 0.4
#> [242,] 0.6 0.4
#> [243,] 0.3 0.7
#> [244,] 0.6 0.4
#> [245,] 0.3 0.7
#> [246,] 0.3 0.7
#> [247,] 0.6 0.4
#> [248,] 0.3 0.7
#> [249,] 0.3 0.7
#> [250,] 0.6 0.4
#> [251,] 0.6 0.4
#> [252,] 0.6 0.4
#> [253,] 0.6 0.4
#> [254,] 0.6 0.4
#> [255,] 0.3 0.7
#> [256,] 0.6 0.4
#> [257,] 0.3 0.7
#> [258,] 0.6 0.4
#> [259,] 0.3 0.7
#> [260,] 0.3 0.7
#> [261,] 0.6 0.4
#> [262,] 0.6 0.4
#> [263,] 0.3 0.7
#> [264,] 0.3 0.7
#> [265,] 0.3 0.7
#> [266,] 0.3 0.7
#> [267,] 0.3 0.7
#> [268,] 0.3 0.7
#> [269,] 0.3 0.7
#> [270,] 0.3 0.7
#> [271,] 0.3 0.7
#> [272,] 0.6 0.4
#> [273,] 0.3 0.7
#> [274,] 0.6 0.4
#> [275,] 0.3 0.7
#> [276,] 0.6 0.4
#> [277,] 0.6 0.4
#> [278,] 0.3 0.7
#> [279,] 0.3 0.7
#> [280,] 0.6 0.4
#> [281,] 0.3 0.7
#> [282,] 0.6 0.4
#> [283,] 0.3 0.7
#> [284,] 0.3 0.7
#> [285,] 0.6 0.4
#> [286,] 0.6 0.4
#> [287,] 0.6 0.4
#> [288,] 0.3 0.7
#> [289,] 0.3 0.7
#> [290,] 0.3 0.7
#> [291,] 0.3 0.7
#> [292,] 0.6 0.4
#> [293,] 0.6 0.4
#> [294,] 0.3 0.7
#> [295,] 0.3 0.7
#> [296,] 0.6 0.4
#> [297,] 0.6 0.4
#> [298,] 0.3 0.7
#> [299,] 0.6 0.4
#> [300,] 0.3 0.7
#> [301,] 0.3 0.7
#> [302,] 0.3 0.7
#> [303,] 0.3 0.7
#> [304,] 0.3 0.7
#> [305,] 0.6 0.4
#> [306,] 0.6 0.4
#> [307,] 0.6 0.4
#> [308,] 0.6 0.4
#> [309,] 0.3 0.7
#> [310,] 0.6 0.4
#> [311,] 0.3 0.7
#> [312,] 0.3 0.7
#> [313,] 0.6 0.4
#> [314,] 0.6 0.4
#> [315,] 0.6 0.4
#> [316,] 0.3 0.7
#> [317,] 0.3 0.7
#> [318,] 0.3 0.7
#> [319,] 0.6 0.4
#> [320,] 0.3 0.7
#> [321,] 0.3 0.7
#> [322,] 0.3 0.7
#> [323,] 0.6 0.4
#> [324,] 0.6 0.4
#> [325,] 0.3 0.7
#> [326,] 0.6 0.4
#> [327,] 0.6 0.4
#> [328,] 0.3 0.7
#> [329,] 0.6 0.4
#> [330,] 0.6 0.4
#> [331,] 0.6 0.4
#> [332,] 0.3 0.7
#> [333,] 0.3 0.7
#> [334,] 0.3 0.7
#> [335,] 0.6 0.4
#> [336,] 0.3 0.7
#> [337,] 0.3 0.7
#> [338,] 0.6 0.4
#> [339,] 0.6 0.4
#> [340,] 0.3 0.7
#> [341,] 0.3 0.7
#> [342,] 0.6 0.4
#> [343,] 0.3 0.7
#> [344,] 0.3 0.7
#> [345,] 0.6 0.4
#> [346,] 0.3 0.7
#> [347,] 0.6 0.4
#> [348,] 0.6 0.4
#> [349,] 0.6 0.4
#> [350,] 0.3 0.7
#> [351,] 0.6 0.4
#> [352,] 0.3 0.7
#> [353,] 0.3 0.7
#> [354,] 0.6 0.4
#> [355,] 0.3 0.7
#> [356,] 0.6 0.4
#> [357,] 0.6 0.4
#> [358,] 0.3 0.7
#> [359,] 0.6 0.4
#> [360,] 0.6 0.4
#> [361,] 0.3 0.7
#> [362,] 0.3 0.7
#> [363,] 0.6 0.4
#> [364,] 0.3 0.7
#> [365,] 0.3 0.7
#> [366,] 0.6 0.4
#> [367,] 0.3 0.7
#> [368,] 0.3 0.7
#> [369,] 0.6 0.4
#> [370,] 0.6 0.4
#> [371,] 0.6 0.4
#> [372,] 0.6 0.4
#> [373,] 0.3 0.7
#> [374,] 0.6 0.4
#> [375,] 0.3 0.7
#> [376,] 0.3 0.7
#> [377,] 0.6 0.4
#> [378,] 0.6 0.4
#> [379,] 0.3 0.7
#> [380,] 0.6 0.4
#> [381,] 0.3 0.7
#> [382,] 0.6 0.4
#> [383,] 0.6 0.4
#> [384,] 0.6 0.4
#> [385,] 0.6 0.4
#> [386,] 0.3 0.7
#> [387,] 0.3 0.7
#> [388,] 0.6 0.4
#> [389,] 0.3 0.7
#> [390,] 0.6 0.4
#> [391,] 0.6 0.4
#> [392,] 0.3 0.7
#> [393,] 0.3 0.7
#> [394,] 0.6 0.4
#> [395,] 0.6 0.4
#> [396,] 0.3 0.7
#> [397,] 0.6 0.4
#> [398,] 0.6 0.4
#> [399,] 0.3 0.7
#> [400,] 0.6 0.4
#> [401,] 0.3 0.7
#> [402,] 0.3 0.7
#> [403,] 0.3 0.7
#> [404,] 0.6 0.4
#> [405,] 0.6 0.4
#> [406,] 0.3 0.7
#> [407,] 0.6 0.4
#> [408,] 0.6 0.4
#> [409,] 0.3 0.7
#> [410,] 0.3 0.7
#> [411,] 0.6 0.4
#> [412,] 0.6 0.4
#> [413,] 0.3 0.7
#> [414,] 0.3 0.7
#> [415,] 0.6 0.4
#> [416,] 0.6 0.4
#> [417,] 0.3 0.7
#> [418,] 0.6 0.4
#> [419,] 0.6 0.4
#> [420,] 0.6 0.4
#> [421,] 0.6 0.4
#> [422,] 0.6 0.4
#> [423,] 0.6 0.4
#> [424,] 0.3 0.7
#> [425,] 0.3 0.7
#> [426,] 0.3 0.7
#> [427,] 0.6 0.4
#> [428,] 0.6 0.4
#> [429,] 0.6 0.4
#> [430,] 0.6 0.4
#> [431,] 0.3 0.7
#> [432,] 0.3 0.7
#> [433,] 0.3 0.7
#> [434,] 0.3 0.7
#> [435,] 0.6 0.4
#> [436,] 0.3 0.7
#> [437,] 0.3 0.7
#> [438,] 0.6 0.4
#> [439,] 0.6 0.4
#> [440,] 0.3 0.7
#> [441,] 0.6 0.4
#> [442,] 0.3 0.7
#> [443,] 0.3 0.7
#> [444,] 0.6 0.4
#> [445,] 0.6 0.4
#> [446,] 0.3 0.7
#> [447,] 0.6 0.4
#> [448,] 0.6 0.4
#> [449,] 0.3 0.7
#> [450,] 0.3 0.7
#> [451,] 0.3 0.7
#> [452,] 0.6 0.4
#> [453,] 0.6 0.4
#> [454,] 0.3 0.7
#> [455,] 0.3 0.7
#> [456,] 0.3 0.7
#> [457,] 0.3 0.7
#> [458,] 0.6 0.4
#> [459,] 0.6 0.4
#> [460,] 0.3 0.7
#> [461,] 0.3 0.7
#> [462,] 0.6 0.4
#> [463,] 0.6 0.4
#> [464,] 0.3 0.7
#> [465,] 0.6 0.4
#> [466,] 0.6 0.4
#> [467,] 0.6 0.4
#> [468,] 0.6 0.4
#> [469,] 0.3 0.7
#> [470,] 0.3 0.7
#> [471,] 0.3 0.7
#> [472,] 0.6 0.4
#> [473,] 0.6 0.4
#> [474,] 0.6 0.4
#> [475,] 0.3 0.7
#> [476,] 0.3 0.7
#> [477,] 0.6 0.4
#> [478,] 0.3 0.7
#> [479,] 0.6 0.4
#> [480,] 0.6 0.4
#> [481,] 0.6 0.4
#> [482,] 0.3 0.7
#> [483,] 0.6 0.4
#> [484,] 0.6 0.4
#> [485,] 0.6 0.4
#> [486,] 0.6 0.4
#> [487,] 0.6 0.4
#> [488,] 0.3 0.7
#> [489,] 0.3 0.7
#> [490,] 0.3 0.7
#> [491,] 0.6 0.4
#> [492,] 0.3 0.7
#> [493,] 0.6 0.4
#> [494,] 0.3 0.7
#> [495,] 0.6 0.4
#> [496,] 0.6 0.4
#> [497,] 0.3 0.7
#> [498,] 0.6 0.4
#> [499,] 0.3 0.7
#> [500,] 0.3 0.7
#> 
#> 
#> $pEvidence
#>   [1] 0.4851495 0.4851495 0.4410450 0.4410450 0.4410450 0.4851495 0.4410450
#>   [8] 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495 0.0490050 0.4851495
#>  [15] 0.4410450 0.4410450 0.4851495 0.4410450 0.4851495 0.4851495 0.4851495
#>  [22] 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450 0.4410450 0.4851495
#>  [29] 0.4851495 0.4410450 0.4410450 0.4410450 0.4851495 0.4851495 0.4851495
#>  [36] 0.4410450 0.4851495 0.0049005 0.4851495 0.4410450 0.4410450 0.4851495
#>  [43] 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495
#>  [50] 0.4851495 0.4851495 0.4410450 0.4851495 0.4410450 0.4851495 0.4851495
#>  [57] 0.4410450 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495
#>  [64] 0.4410450 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495
#>  [71] 0.4410450 0.4851495 0.4410450 0.4851495 0.4851495 0.4851495 0.0049005
#>  [78] 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495
#>  [85] 0.0049005 0.4410450 0.4410450 0.4410450 0.4410450 0.4851495 0.4410450
#>  [92] 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495 0.4851495 0.4851495
#>  [99] 0.4851495 0.4410450 0.0490050 0.4410450 0.4410450 0.4410450 0.4410450
#> [106] 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495
#> [113] 0.4851495 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495
#> [120] 0.4851495 0.4410450 0.4851495 0.4410450 0.4851495 0.4851495 0.4410450
#> [127] 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450
#> [134] 0.4851495 0.0490050 0.4410450 0.0490050 0.4851495 0.4410450 0.4410450
#> [141] 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495 0.4851495 0.4410450
#> [148] 0.0049005 0.4410450 0.4851495 0.4851495 0.4851495 0.4410450 0.4851495
#> [155] 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495
#> [162] 0.4410450 0.4410450 0.4851495 0.4410450 0.4851495 0.4851495 0.4851495
#> [169] 0.4410450 0.0490050 0.0000495 0.4851495 0.4410450 0.4851495 0.4410450
#> [176] 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450 0.4851495
#> [183] 0.4410450 0.4410450 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450
#> [190] 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495
#> [197] 0.4410450 0.4410450 0.4851495 0.4851495 0.4851495 0.4851495 0.0490050
#> [204] 0.4851495 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495
#> [211] 0.4410450 0.4410450 0.4851495 0.0490050 0.4851495 0.4410450 0.4410450
#> [218] 0.4851495 0.4851495 0.4851495 0.4410450 0.4851495 0.4410450 0.4851495
#> [225] 0.4410450 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450
#> [232] 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495
#> [239] 0.4851495 0.4851495 0.4410450 0.4410450 0.4851495 0.4410450 0.4851495
#> [246] 0.4851495 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450 0.4410450
#> [253] 0.4410450 0.4410450 0.4851495 0.4410450 0.4851495 0.4410450 0.4851495
#> [260] 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495 0.4851495 0.4851495
#> [267] 0.4851495 0.4851495 0.4851495 0.4851495 0.4851495 0.0490050 0.4851495
#> [274] 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495 0.0490050
#> [281] 0.4851495 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450 0.4410450
#> [288] 0.4851495 0.4851495 0.4851495 0.4851495 0.4410450 0.4410450 0.4851495
#> [295] 0.4851495 0.4410450 0.4410450 0.4851495 0.4410450 0.4851495 0.0049005
#> [302] 0.4851495 0.4851495 0.4851495 0.4410450 0.4410450 0.0490050 0.4410450
#> [309] 0.4851495 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450 0.4410450
#> [316] 0.4851495 0.4851495 0.4851495 0.4410450 0.4851495 0.4851495 0.4851495
#> [323] 0.4410450 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495 0.4410450
#> [330] 0.4410450 0.4410450 0.4851495 0.4851495 0.0049005 0.4410450 0.4851495
#> [337] 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495
#> [344] 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4410450 0.4851495
#> [351] 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450
#> [358] 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450 0.4851495
#> [365] 0.4851495 0.4410450 0.4851495 0.4851495 0.4410450 0.4410450 0.4410450
#> [372] 0.4410450 0.4851495 0.0490050 0.4851495 0.4851495 0.4410450 0.0490050
#> [379] 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4410450 0.4410450
#> [386] 0.4851495 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495
#> [393] 0.4851495 0.4410450 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495
#> [400] 0.4410450 0.4851495 0.4851495 0.4851495 0.4410450 0.0490050 0.4851495
#> [407] 0.0042750 0.4410450 0.4851495 0.4851495 0.4410450 0.0490050 0.4851495
#> [414] 0.4851495 0.0042750 0.4410450 0.4851495 0.4410450 0.4410450 0.4410450
#> [421] 0.4410450 0.4410450 0.4410450 0.4851495 0.4851495 0.4851495 0.4410450
#> [428] 0.4410450 0.0490050 0.4410450 0.4851495 0.4851495 0.4851495 0.4851495
#> [435] 0.4410450 0.4851495 0.4851495 0.0490050 0.4410450 0.4851495 0.4410450
#> [442] 0.4851495 0.4851495 0.4410450 0.0490050 0.4851495 0.4410450 0.4410450
#> [449] 0.4851495 0.4851495 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495
#> [456] 0.0049005 0.4851495 0.4410450 0.4410450 0.4851495 0.4851495 0.4410450
#> [463] 0.4410450 0.4851495 0.4410450 0.4410450 0.4410450 0.4410450 0.4851495
#> [470] 0.4851495 0.0049005 0.0490050 0.4410450 0.4410450 0.4851495 0.4851495
#> [477] 0.0002250 0.4851495 0.4410450 0.4410450 0.4410450 0.4851495 0.4410450
#> [484] 0.0490050 0.4410450 0.4410450 0.4410450 0.4851495 0.4851495 0.4851495
#> [491] 0.4410450 0.4851495 0.4410450 0.4851495 0.4410450 0.4410450 0.4851495
#> [498] 0.4410450 0.4851495 0.4851495
#>