loadHuginNet(gRain)	R Documentation

Load and save Hugin net files

Description

These functions can load a net file saved in the 'Hugin format' into R and save a network in R as a file in the 'Hugin format'.

Usage


loadHuginNet(file, description, details = 0)
saveHuginNet(gin, file, details = 0)

Arguments

gin 	 An independence network

file 	 Name of HUGIN net file. Convenient to give the file the extension '.net'

description 	 A text describing the network, defaults to file

details 	 Debugging information

Value

An object (a list) of class "huginNet".

Author(s)

Søren Højsgaard, sorenh@agrsci.dk

See Also

grain

Examples



tf <- system.file("huginex", "chest_clinic.net", package = "gRain")
chest <- loadHuginNet(tf, details=1)
chest 

td <- tempdir()
saveHuginNet(chest, paste(td,"/chest.net",sep=''))

chest2 <- loadHuginNet(paste(td,"/chest.net",sep=''))

tf <- system.file("huginex", "golf.net", package = "gRain")
golf <- loadHuginNet(tf, details=1)

saveHuginNet(golf, paste(td,"/golf.net",sep=''))
golf2 <- loadHuginNet(paste(td,"/golf.net",sep=''))


[Package gRain version 0.8.2 Index]