propagate.grain(gRain)	R Documentation

Propagate a graphical independence network (a Bayesian network)

Description

Propagation refers to calibrating the cliques of the junction tree so that the clique potentials are consistent on their intersections

Usage


propagate.grain(object, details = object$details,...)

Arguments

object 	 A grain object

details 	 For debugging info

... 	 Currently not used

Value

A compiled and propagated grain object.

Author(s)

Søren Højsgaard, sorenh@agrsci.dk

See Also

grain,compile

[Package gRain version 0.8.2 Index]